Make control hints context-aware
Also improve layout and refactor to avoid repeated code.

Close #1449
